fastapi_revision/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy

# hashing
passlib
bcrypt

# foe JWT token
python-jose